refactor: removendo código da feature/finance, movendo RealizarPagamento para feature delivery
