+++ frontend/src/app/routes.tsx
@@ -6,7 +6,7 @@ import { clienteRoutes } from "../features/clients/routes/cliente.routes";
 import { usuarioRoutes } from "../features/users/routes/usuario.routes";
 import { adminRoutes } from "../features/admin/routes/admin.routes";
 import { homeRoutes } from "../features/home/routes/home.routes";
-import { financeRoutes } from "@/features/finance/routes/finance.routes";
+import { deliveryRoutes } from "@/features/services/delivery/routes/delivery.routes";
 import { executionRoutes } from "@/features/services/execution/routes/execution.routes";
 
 const AppRoutes = () => {
@@ -18,7 +18,7 @@ const AppRoutes = () => {
     ...servicoRoutes,
     ...executionRoutes,
     ...adminRoutes,
-    ...financeRoutes,
+    ...deliveryRoutes,
   ];
 
   return (
