--- frontend/src/features/finance/routes/finance.routes.tsx
@@ -1,6 +1,0 @@
-import RealizarPagamento from "../pages/RealizarPagamento";
-import { financePaths } from "./financePaths";
-
-export const financeRoutes = [
-  { path: financePaths.realizarPagamento, element: <RealizarPagamento /> },
-];
