+++ frontend/src/features/freelancers/validations/contaBancaria.schema.ts
@@ -27,16 +27,3 @@ export const BankAccountSchema = yup.object({
     .required("Selecione o tipo da conta")
     .oneOf([1, 2], "Tipo de conta inválido"),
 });
-
-export type ICardForm = yup.InferType<typeof CardSchema>;
-
-export const CardSchema = yup.object({
-  name: yup.string().required("Nome do impresso é obrigatório"),
-  cvv: yup
-    .string()
-    .required("CVV é obrigatório")
-    .min(3, "CVV inválido")
-    .max(3, "CVV inválido"),
-
-  flag: yup.number().required("Bandeira é obrigatória"),
-});
