+++ frontend/src/features/services/delivery/routes/delivery.routes.tsx
@@ -1,0 +1,6 @@
+import RealizarPagamento from "../pages/RealizarPagamento";
+import { deliveryPaths } from "./deliveryPaths";
+
+export const deliveryRoutes = [
+  { path: deliveryPaths.realizarPagamento, element: <RealizarPagamento /> },
+];
