+++ frontend/src/features/services/delivery/routes/deliveryPaths.ts
@@ -1,4 +1,4 @@
-export const financePaths = {
+export const deliveryPaths = {
   realizarPagamento: "/realizar-pagamento/:id",
 
   realizarPagamentoById: (id: string | number) => `/realizar-pagamento/${id}`,
